end of section 18

+++ client/src/app/messages/messages.component.ts
@@ -1,9 +1,11 @@
-import { Component, OnInit } from '@angular/core';
+import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
+  changeDetection: ChangeDetectionStrategy.OnPush,
   selector: 'app-messages',
   templateUrl: './messages.component.html',
   styleUrls: ['./messages.component.css']
@@ -16,7 +18,7 @@ export class MessagesComponent implements OnInit {
   pageSize=5;
   loading=false;
 
-  constructor(private messageService:MessageService) { }
+  constructor(private messageService:MessageService,private confirmService:ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -32,9 +34,14 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id:number){
-     this.messageService.deleteMessage(id).subscribe(()=> {
-      this.messages.splice(this.messages.findIndex(m=>m.id===id),1);
-     })
+      this.confirmService.confirm('Confirm Delete Message','This cannot be done').subscribe(result=> {
+        if(result){
+          this.messageService.deleteMessage(id).subscribe(()=> {
+            this.messages.splice(this.messages.findIndex(m=>m.id===id),1);
+           })
+        }
+      })
+     
   }
 
   pageChanged(event:any){
